Move _protocols.py to protocols.py
