+++ lib/galaxy/datatypes/anvio.py
@@ -7,7 +7,7 @@ import logging
 import os
 from typing import Optional
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasExtraFilesAndMetadata,
 )
