+++ lib/galaxy/datatypes/blast.py
@@ -40,7 +40,7 @@ from typing import (
     Optional,
 )
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetHasHidProtocol,
     DatasetProtocol,
 )
