+++ lib/galaxy/datatypes/constructive_solid_geometry.py
@@ -14,7 +14,7 @@ from typing import (
 
 from galaxy import util
 from galaxy.datatypes import data
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasMetadata,
 )
