+++ lib/galaxy/datatypes/gis.py
@@ -1,7 +1,7 @@
 """
 GIS classes
 """
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasExtraFilesAndMetadata,
 )
