+++ lib/galaxy/datatypes/goldenpath.py
@@ -6,7 +6,7 @@ from typing import (
     Union,
 )
 
-from galaxy.datatypes._protocols import DatasetProtocol
+from galaxy.datatypes.protocols import DatasetProtocol
 from galaxy.datatypes.sniff import (
     build_sniff_from_prefix,
     FilePrefix,
