+++ lib/galaxy/datatypes/hdf5.py
@@ -4,7 +4,7 @@ This datatype was created for use with the iSEE interactive tool.
 """
 from typing import Optional
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     HasExtraFilesAndMetadata,
     HasMetadata,
 )
