+++ lib/galaxy/datatypes/images.py
@@ -10,7 +10,7 @@ import mrcfile
 import numpy as np
 import tifffile
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasExtraFilesAndMetadata,
 )
