+++ lib/galaxy/datatypes/microarrays.py
@@ -1,7 +1,7 @@
 import logging
 
 from galaxy.datatypes import data
-from galaxy.datatypes._protocols import DatasetProtocol
+from galaxy.datatypes.protocols import DatasetProtocol
 from galaxy.datatypes.binary import Cel  # noqa: F401
 from galaxy.datatypes.data import get_file_peek
 from galaxy.datatypes.metadata import MetadataElement
