+++ lib/galaxy/datatypes/mothur.py
@@ -8,7 +8,7 @@ from typing import (
     Optional,
 )
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasMetadata,
 )
