+++ lib/galaxy/datatypes/phylip.py
@@ -10,7 +10,7 @@ Phylip datatype sniffer
 from typing import TYPE_CHECKING
 
 from galaxy import util
-from galaxy.datatypes._protocols import DatasetProtocol
+from galaxy.datatypes.protocols import DatasetProtocol
 from galaxy.datatypes.data import (
     get_file_peek,
     Text,
