+++ lib/galaxy/datatypes/plant_tribes.py
@@ -1,7 +1,7 @@
 import logging
 import re
 
-from galaxy.datatypes._protocols import DatasetProtocol
+from galaxy.datatypes.protocols import DatasetProtocol
 from galaxy.datatypes.data import (
     get_file_peek,
     Text,
