+++ lib/galaxy/datatypes/qiime2.py
@@ -11,7 +11,7 @@ from typing import (
 
 import yaml
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasMetadata,
 )
