+++ lib/galaxy/datatypes/qualityscore.py
@@ -3,7 +3,7 @@ Qualityscore class
 """
 import logging
 
-from galaxy.datatypes._protocols import DatasetProtocol
+from galaxy.datatypes.protocols import DatasetProtocol
 from galaxy.datatypes.sniff import (
     build_sniff_from_prefix,
     FilePrefix,
