+++ lib/galaxy/datatypes/registry.py
@@ -20,7 +20,7 @@ from typing import (
 import yaml
 
 import galaxy.util
-from galaxy.datatypes._protocols import DatasetProtocol
+from galaxy.datatypes.protocols import DatasetProtocol
 from galaxy.tool_util.edam_util import load_edam_tree
 from galaxy.util import RW_R__R__
 from galaxy.util.bunch import Bunch
