+++ lib/galaxy/datatypes/sequence.py
@@ -24,7 +24,7 @@ from markupsafe import escape
 
 from galaxy import util
 from galaxy.datatypes import metadata
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetHasHidProtocol,
     DatasetProtocol,
     HasMetadata,
