+++ lib/galaxy/datatypes/tracks.py
@@ -3,7 +3,7 @@ Datatype classes for tracks/track views within galaxy.
 """
 import logging
 
-from galaxy.datatypes._protocols import (
+from galaxy.datatypes.protocols import (
     DatasetProtocol,
     HasExtraFilesAndMetadata,
 )
